move main route builder function to routes
