+++ src/routes/mod.rs
@@ -16,3 +16,6 @@ pub use lines_batch::lines_batch;
 
 mod custom_rejection_handler;
 pub use custom_rejection_handler::custom_rejection_handler;
+
+mod load_data_and_get_combined_routes;
+pub use load_data_and_get_combined_routes::load_data_and_get_combined_routes;
